refactor: move step type in folder
