+++ contexts/export.utils.ts
@@ -1,4 +1,4 @@
-import { OnboardingStep } from "./step.type";
+import { OnboardingStep } from "../OnboardingSteps/step.type";
 
 export const exportSteps = (steps: OnboardingStep[]) => {
   const json = JSON.stringify(steps);
