+++ layout/EditPageForm.tsx
@@ -4,7 +4,7 @@ import { Checkbox, Heading, Label, TextArea, XStack, YStack } from "tamagui";
 import { Check as CheckIcon } from '@tamagui/lucide-icons'
 import React from "react";
 import { SelectType } from "../components/SelectType";
-import { getInitialStepPayload, OnboardingStep, StepType, STEP_TYPES } from "../contexts/step.type";
+import { getInitialStepPayload, OnboardingStep, StepType, STEP_TYPES } from "../OnboardingSteps/step.type";
 import { MediaContentEditor } from "../OnboardingSteps/MediaContent/MediaContent.form";
 import { NativeSyntheticEvent, TextInputChangeEventData } from "react-native";
 import { QuestionEditor } from "../OnboardingSteps/Question/Question.form";
