+++ layout/StepsList.tsx
@@ -8,7 +8,7 @@ import { HTML5Backend } from 'react-dnd-html5-backend'
 import { Button, RadioGroup, ScrollView, View, YStack } from 'tamagui'
 import { useSteps } from '../contexts/steps-context'
 import { v4 as uuidv4 } from 'uuid'
-import { QuestionStepType, OnboardingStep, STEP_TYPES } from '../contexts/step.type'
+import { QuestionStepType, OnboardingStep, STEP_TYPES } from '../OnboardingSteps/step.type'
 import { ExportSheet } from './ExportSheet'
 import React from 'react'
 import { ImportSheet } from './ImportSheet'
